requirements.txt
semver==3.0.2
folium==0.18.0
h3==3.7.7
pybloomfiltermmap3==0.5.5
PyYAML==6.0
git+https://github.com/databrickslabs/geoscan.git#subdirectory=python